[WIP] Compress raw payload
Compress raw payload in the dynamic product details table.  Since this
table is the one that is going to grow every week or so, it's worth
reducing its size if we can.  You can read issue #6 for some analysis.

I also checked the size reduction on an actual inventory database, not
just on the raw payload data itself.  I used a sample database with 901
products.  Once scrape-prices run, which adds 901 samples, adds 503808
bytes to the database without compression, and 139264 bytes with
compression.  Extrapolating to 120000 products, that means about 64 MiB
vs 18 MiB.

As explained in issue #6, one factor that helps compression is the
use of a pre-computed dictionary.  I used pyzstd's train_dict function
[1] on all the raw payloads of one full run (which means about 120000
entries).  I saved that dictionary and embedded it as bytes in a Python
file (that seemed simpler than try to publish it as a separate data file
of the Python package).  I encoded all bytes using the \xNN notation
(even alphanumeric bytes), because it contains a lot of Canadian Tire
"specific" words coming from item descriptions, and I don't want these
words to be indexed and found by search engines.  You can display the
"raw" dictionary using:

    $ poetry run python zstd_dict.py
    $ poetry run python zstd_dict.py > foo
    $ file foo
    foo: Zstandard dictionary (ID 1593333190)

Modify storage.py to:

  - instantiate a Zstd compressor using that dictionary
  - change the raw_payload column type to LargeBinary
  - pass the raw payload through the compressor when saving it.  Use the
    FLUSH_FRAME mode, because we re-use the compressor, but want each
    compression to be independent.

[1] https://pyzstd.readthedocs.io/en/latest/#train_dict

Closes: https://github.com/jgalar/CanadianTracker/issues/6