Split scrape-inventory into scrape-products + scrape-skus
We currently attempt to get our list of SKUs, for each product, from the
request we use to list products themselves.  The json objects for
products used to include a `skus` list, from which we could get SKU
codes (and nothing more).  The responses today still have a `skus`
attribute, but the list is now empty, meaning that we don't discover new
SKUs.

The only obvious way I see to obtain the SKU list for all products is to
fetch the data that is fetched by the CT website when viewing a product
page.

The downside of this is that it requires a lot more requests to obtain
the SKU list of all products: one request per product.  We have about
132k products today.  A trial run took about 24h to complete.

The upside of this is that this will allow us to save more information
on each SKU, like what their specificities are, which will allow us to
improve our SKU list displays.

To implement this, split the scrape-inventory command in scrape-products
and scrape-skus.  Having two commands allows to keep each step separate,
which makes it easier to implement and debug.