Add one functional test
Add a pytest fixture that will be useful to write tests, and add a
simple test to exercise it.  The fixture allows injecting test functions
with a temporary database initialized with the current SQLAlchemy
schema.

I had to modify a past migration, otherwise it wouldn't work on the CI.
Not sure why it works locally, maybe an sqlite version thing.  I don't
really have time to dig more.